Implemented lazy sessions and persistence
`SessionPersistenceInterface` describes methods for creating a
`SessionInterface` instance, as well as persisting it (along with
providing an altered `ResponseInterface` to return to the client).
Additionally, it provides both a static and an instance method for
generating instances of itself; this allows usage with persistence
mechanisms that have no dependencies (ext-session, JWT), as well as
those that do (usage of redis, memcached, etc).

`SessionMiddleware` now composes either a class name or implementation
of `SessionPersistenceInterface`, and uses the appropriate factory
method on that value to create an instance. This is passed to a new
class, `LazySession`, a `SessionInterface` implementation. This class
will use the persistence instance's `createSession()` method the first
time any method is called on it, and then proxy to it. This allows
delaying session initialization and session data parsing until it is
actually needed. Finally, after receiving a response, the middleware
passes the `LazySession` to the persistence instance's
`persistSession()` method (along with the response), in order to return
a response.