Add identifier awareness to sessions
In working with Swoole, I have discovered that services simply cannot be
stateful: changing state in a service composed in the dependency
injection container has ramifications for any concurrent or subsequent
requests.

One area I witnessed this was with zend-expressive-session, specifically
with persistence implementations. In order to perform persistence, they
needed access to the original session identifier (either to store data,
or to remove a previous session when regenerating). Because session
instances were not aware of their identifier, this meant the persistence
implementations had to either:

- store the identifier as a property
- add a "magic" value to the session instance

The former leads to the state issues, while the latter is a hack.

This patch adds a new interface, `SessionIdentifierAwareInterface`,
which can be composed in `SessionInterface` implementations in order to
provide access to the identifier. The default `Session` implementation
now accepts an optional identifier to its constructor, and exposes it
via that method. The `LazySession` implementation checks to see if the
proxied session implements the interface, and, if so, proxies to it.

This approach allows engines to create instances that implement the new
interface and store the identifier discretely as part of the instance.
They can then query for the value when it comes time to persist data.